@@ -2,27 +2,30 @@ name: zpevnik
 
 publish_to: 'none'
 
-version: 2.6.0+32
+version: 3.0.0+104
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  sqflite: ^2.0.0+3
-  sqfentity: ^2.1.2+1
-  sqfentity_gen: ^2.0.0
+  sqflite: ^2.0.2+1
   path: ^1.8.0
+
+  objectbox: ^1.5.0
+  objectbox_flutter_libs: any
 
   shared_preferences: ^2.0.6
   intl: ^0.17.0
   connectivity: ^3.0.6
   diacritic: ^0.1.3
   wakelock: ^0.6.1+2
-    
+
   provider: ^6.0.0
+
+  pool: ^1.5.0
 
   modal_bottom_sheet: ^2.0.0
 
@@ -30,26 +33,31 @@ dependencies:
   share_plus: ^4.0.3
 
   youtube_player_flutter: ^8.0.0
-  miniplayer: ^1.0.1
-  
+  just_audio: ^0.9.27
+
+  flutter_cached_pdfview: ^0.4.1
+
   flutter_svg: ^1.0.0
-  flutter_reorderable_list: ^1.1.0
   flutter_staggered_grid_view: ^0.6.1
 
-  scroll_to_index: ^2.1.1
+  adaptive_dialog: ^1.6.4
+
+  flutter_speed_dial: ^6.0.0
 
   # firebase_analytics: ^9.0.4
+  sign_in_with_apple: ^4.0.0
 
   uni_links: ^0.5.1
 
-  # flutter_nearby_connections: ^1.0.11
   device_info: ^2.0.2
 
   auto_size_text: ^3.0.0-nullsafety.0
 
-  fluttertoast: ^8.0.8
+  flutter_quill: ^5.0.4
 
-  flutter_slidable: ^1.1.0
+  dotted_border: ^2.0.0+2
+
+  font_awesome_flutter: ^10.1.0
 
   cupertino_icons: ^1.0.3
 
@@ -57,6 +65,7 @@ dependencies:
 
 dev_dependencies:
   build_runner: ^2.0.4
+  objectbox_generator: any
 
   flutter_lints: ^1.0.0
 
@@ -83,5 +92,9 @@ flutter:
   assets:
     - assets/data.json
     - assets/images/
+    - assets/images/logos/
     - assets/images/songbooks/
     - assets/
+
+objectbox:
+  output_dir: models
