@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
 
-const int kCurrentVersion = 8;
-
-const double kDefaultPadding = 16;
-const double kDefaultRadius = 10;
+const double kDefaultPadding = 12;
+const double kDefaultRadius = 12;
 const kDefaultAnimationDuration = Duration(milliseconds: 300);
 
 const double kMinimumFontSizeScale = 0.5;
