@@ -2,7 +2,9 @@ const proscholyUrl = 'https://zpevnik.proscholy.cz';
 
 const songUrl = '$proscholyUrl/pisen';
 
-const reportUrl = 'https://proscholy.atlassian.net/servicedesk/customer/portal/1/group/1/create/19';
+const reportSongLyricUrl = 'https://proscholy.atlassian.net/servicedesk/customer/portal/1/group/1/create/19';
+
+const reportUrl = 'https://glowspace.atlassian.net/servicedesk/customer/portal/1/group/6/create/20';
 
 const feedbackAndroidUrl =
     'https://docs.google.com/forms/d/e/1FAIpQLSfI0143gkLBtMbWQnSa9nzpOoBNMokZrOIS5mUreSR41E_B7A/viewform?usp=pp_url&entry.1865829262=ano,+verzi+pro+Android';
