@@ -1,32 +1,41 @@
-import 'package:zpevnik/models/model.dart' as model;
+// ignore: unnecessary_import
+import 'package:objectbox/objectbox.dart';
+import 'package:zpevnik/models/objectbox.g.dart';
+import 'package:zpevnik/models/song_lyric.dart';
+import 'package:zpevnik/models/playlist.dart';
 
-// wrapper around PlaylistRecord db model for easier field access
-class PlaylistRecord {
-  final model.PlaylistRecord entity;
+@Entity()
+class PlaylistRecord implements Comparable<PlaylistRecord> {
+  int id = 0;
 
-  PlaylistRecord(this.entity);
+  int rank;
 
-  static Future<PlaylistRecord> create(int songLyricId, int playlistId, int rank) async {
-    final entity = model.PlaylistRecord(song_lyricsId: songLyricId, playlistsId: playlistId, rank: rank);
+  final songLyric = ToOne<SongLyric>();
+  final playlist = ToOne<Playlist>();
 
-    entity.id = await entity.save();
+  PlaylistRecord(this.rank);
 
-    return PlaylistRecord(entity);
+  static int nextRank(Store store, Playlist playlist) {
+    final query = store.box<PlaylistRecord>().query(PlaylistRecord_.playlist.equals(playlist.id));
+    query.order(PlaylistRecord_.rank);
+
+    final rank = query.build().findFirst()?.rank ?? -1;
+
+    return rank + 1;
   }
 
-  static Future<List<PlaylistRecord>> get playlistRecords async {
-    final entities = await model.PlaylistRecord().select().toList();
+  PlaylistRecord copyWith({Playlist? playlist}) {
+    final playlistRecord = PlaylistRecord(rank)
+      ..playlist.target = playlist
+      ..songLyric.target = songLyric.target;
 
-    return entities.map((entity) => PlaylistRecord(entity)).toList();
+    return playlistRecord;
   }
 
-  int get id => entity.id ?? 0;
-  int get playlistId => entity.playlistsId ?? 0;
-  int get songLyricId => entity.song_lyricsId ?? 0;
-  int get rank => entity.rank ?? 0;
+  @override
+  String toString() =>
+      'PlaylistRecord(id: $id, rank: $rank, songLyric: ${songLyric.targetId}, playlist: ${playlist.targetId})';
 
-  set rank(int value) {
-    entity.rank = value;
-    entity.save();
-  }
+  @override
+  int compareTo(PlaylistRecord other) => rank.compareTo(other.rank);
 }
