@@ -1,13 +1,13 @@
-import 'dart:math';
-
-import 'package:flutter/material.dart';
-import 'package:zpevnik/constants.dart';
+// ignore: unnecessary_import
+import 'package:objectbox/objectbox.dart';
 import 'package:zpevnik/models/author.dart';
 import 'package:zpevnik/models/external.dart';
-import 'package:zpevnik/models/model.dart' as model;
-import 'package:zpevnik/models/songbook.dart';
+import 'package:zpevnik/models/objectbox.g.dart';
+import 'package:zpevnik/models/playlist.dart';
+import 'package:zpevnik/models/playlist_record.dart';
+import 'package:zpevnik/models/song.dart';
 import 'package:zpevnik/models/songbook_record.dart';
-import 'package:zpevnik/providers/data.dart';
+import 'package:zpevnik/models/tag.dart';
 
 enum SongLyricType {
   original,
@@ -16,7 +16,7 @@ enum SongLyricType {
 }
 
 extension SongLyricTypeExtension on SongLyricType {
-  static SongLyricType fromString(String string) {
+  static SongLyricType fromString(String? string) {
     switch (string) {
       case "ORIGINAL":
         return SongLyricType.original;
@@ -27,7 +27,27 @@ extension SongLyricTypeExtension on SongLyricType {
     }
   }
 
-  int get rawValue => SongLyricType.values.indexOf(this);
+  static SongLyricType fromRawValue(int rawValue) {
+    switch (rawValue) {
+      case 0:
+        return SongLyricType.original;
+      case 1:
+        return SongLyricType.authorizedTranslation;
+      default:
+        return SongLyricType.translation;
+    }
+  }
+
+  int get rawValue {
+    switch (this) {
+      case SongLyricType.original:
+        return 0;
+      case SongLyricType.authorizedTranslation:
+        return 1;
+      default:
+        return 2;
+    }
+  }
 
   String get description {
     switch (this) {
@@ -39,146 +59,150 @@ extension SongLyricTypeExtension on SongLyricType {
         return "Překlad";
     }
   }
-
-  Color get color {
-    switch (this) {
-      case SongLyricType.original:
-        return blue;
-      case SongLyricType.authorizedTranslation:
-        return yellow;
-      default:
-        return green;
-    }
-  }
 }
 
-// wrapper around SongLyric db model for easier field access
+@Entity()
 class SongLyric {
-  final model.SongLyric entity;
+  @Id(assignable: true)
+  final int id;
 
-  static int _nextFavoriteOrder = 0;
+  final String name;
+  final String? secondaryName1;
+  final String? secondaryName2;
 
-  SongLyric(this.entity) {
-    _nextFavoriteOrder = max(_nextFavoriteOrder, favoriteRank + 1);
+  final String? lyrics;
+  final String? lilypond;
+
+  final String? lang;
+  final String? langDescription;
+
+  final int dbType;
+  final bool hasChords;
+
+  int? accidentals;
+  bool? showChords;
+  int transposition = 0;
+
+  final authors = ToMany<Author>();
+  final song = ToOne<Song>();
+  final tags = ToMany<Tag>();
+
+  @Backlink()
+  final externals = ToMany<External>();
+
+  @Backlink()
+  final songbookRecords = ToMany<SongbookRecord>();
+
+  @Backlink()
+  final playlistRecords = ToMany<PlaylistRecord>();
+
+  SongLyric(
+    this.id,
+    this.name,
+    this.secondaryName1,
+    this.secondaryName2,
+    this.lyrics,
+    this.lilypond,
+    this.lang,
+    this.langDescription,
+    this.dbType,
+    this.hasChords,
+  );
+
+  factory SongLyric.fromJson(Map<String, dynamic> json, Store store) {
+    final id = int.parse(json['id'] as String);
+
+    final authors = store.box<Author>().getMany(
+        (json['authors_pivot'] as List).map((json) => int.parse(json['pivot']['author']['id'] as String)).toList());
+    final tags =
+        store.box<Tag>().getMany((json['tags'] as List).map((json) => int.parse(json['id'] as String)).toList());
+
+    final songbookRecords = SongbookRecord.fromMapList(json, id);
+    final externals = External.fromMapList(json, id);
+
+    return SongLyric(
+      id,
+      json['name'] as String,
+      json['secondary_name_1'] as String?,
+      json['secondary_name_2'] as String?,
+      json['lyrics'] as String?,
+      json['lilypond_svg'] as String?,
+      json['lang'] as String,
+      json['lang_string'] as String,
+      SongLyricTypeExtension.fromString(json['type_enum'] as String?).rawValue,
+      json['has_chords'] as bool,
+    )
+      ..authors.addAll(authors.cast())
+      ..song.targetId = json['song'] == null ? null : int.parse(json['song']['id'] as String)
+      ..tags.addAll(tags.cast())
+      ..externals.addAll(externals)
+      ..songbookRecords.addAll(songbookRecords);
   }
 
-  static Future<List<SongLyric>> get songLyrics async {
-    final entities =
-        await model.SongLyric().select().lyrics.not.isNull().or.lilypond_svg.not.isNull().orderBy('name').toList();
-
-    return entities.map((entity) => SongLyric(entity)).toList();
+  static List<SongLyric> fromMapList(Map<String, dynamic> json, Store store) {
+    return (json['song_lyrics'] as List).map((json) => SongLyric.fromJson(json, store)).toList();
   }
 
-  final Map<int, SongbookRecord> songbookRecords = {};
+  static List<SongLyric> load(Store store) {
+    final query = store.box<SongLyric>().query(SongLyric_.lyrics.notNull().or(SongLyric_.lilypond.notNull()));
+    query.order(SongLyric_.name);
 
-  int get id => entity.id ?? 0;
-  String get name => entity.name ?? '';
-  String get lyrics => entity.lyrics ?? '';
-  String? get lilypond => entity.lilypond_svg;
-  String get language => entity.lang_string ?? '';
-  SongLyricType get type => SongLyricTypeExtension.fromString(entity.type_enum ?? '');
-  int? get songId => entity.songsId;
-
-  int get favoriteRank => entity.favorite_rank ?? -1;
-  bool get isFavorite => entity.favorite_rank != null;
-
-  bool? get showChords => entity.show_chords;
-  int get transposition => entity.transposition ?? 0;
-  int? get accidentals => entity.accidentals;
-
-  // TODO: this should not be here
-  Key get key => Key(id.toString());
-
-  String? get secondaryName {
-    String? name;
-
-    if (entity.secondary_name_1 != null) name = '${entity.secondary_name_1}';
-    if (entity.secondary_name_2 != null) name = '$name\n${entity.secondary_name_2}';
-
-    return name;
+    return query.build().find();
   }
 
-  set favoriteRank(int value) {
-    entity.favorite_rank = value;
-    entity.save();
-  }
+  SongLyricType get type => SongLyricTypeExtension.fromRawValue(dbType);
 
-  set showChords(bool? newValue) {
-    entity.show_chords = newValue;
-    entity.save();
-  }
+  bool get hasTranslations => (song.target?.songLyrics.length ?? 1) > 1;
 
-  set transposition(int newValue) {
-    if (newValue == 12 || newValue == -12) newValue = 0;
+  bool get hasLyrics => lyrics != null && lyrics!.isNotEmpty;
+  bool get hasFiles => externals.any((external) => external.type == MediaType.pdf);
+  bool get hasRecordings => externals.any((external) => external.type == MediaType.youtube);
 
-    entity.transposition = newValue;
-    entity.save();
-  }
+  bool get isFavorite =>
+      playlistRecords.any((playlistRecord) => playlistRecord.playlist.targetId == favoritesPlaylistId);
 
-  set accidentals(int? newValue) {
-    entity.accidentals = newValue;
-    entity.save();
-  }
-
-  void toggleFavorite() {
-    if (isFavorite) {
-      entity.favorite_rank = null;
-    } else {
-      entity.favorite_rank = _nextFavoriteOrder++;
-    }
-
-    entity.save();
-  }
-
-  String number(Songbook? songbook) {
-    if (songbook != null) return songbookRecords[songbook.id]?.number ?? '0';
-
-    return entity.id.toString();
-  }
-
-  final List<int> tagIds = [];
-  final List<Author> authors = [];
-
-  String? _authorsText;
-  String authorsText(DataProvider provider) {
-    if (_authorsText != null) return _authorsText!;
-
-    String text;
-
+  String get authorsText {
     if (type == SongLyricType.original) {
-      if (authors.length == 0)
-        text = 'Autor neznámý';
-      else if (authors.length == 1)
-        text = 'Autor: ${authors[0].name}';
-      else
-        text = 'Autoři: ${authors.map((author) => author.name).toList().join(", ")}';
+      if (authors.isEmpty) {
+        return 'Autor neznámý';
+      } else if (authors.length == 1) {
+        return 'Autor: ${authors[0].name}';
+      } else {
+        return 'Autoři: ${authors.map((author) => author.name).toList().join(", ")}';
+      }
     } else {
       String originalText = '';
 
-      try {
-        final songLyrics = provider.songsSongLyrics(songId ?? -1);
-        final original = songLyrics?.firstWhere((songLyric) => songLyric.type == SongLyricType.original);
+      final original = song.target?.songLyrics
+          .cast()
+          .firstWhere((songLyric) => songLyric.type == SongLyricType.original, orElse: () => null);
 
-        if (original != null) {
-          originalText = 'Originál: ${original.name}\n';
+      if (original != null) {
+        originalText = 'Originál: ${original.name}\n';
 
-          originalText += '${original.authorsText(provider)}\n';
-        }
-      } on StateError {}
+        originalText += '${original.authorsText}\n';
+      }
 
-      if (authors.length == 0)
-        text = '${originalText}Autor překladu neznámý';
-      else if (authors.length == 1)
-        text = '${originalText}Autor překladu: ${authors[0].name}';
-      else
-        text = '${originalText}Autoři překladu: ${authors.map((author) => author.name).toList().join(", ")}';
+      if (authors.isEmpty) {
+        return '${originalText}Autor překladu neznámý';
+      } else if (authors.length == 1) {
+        return '${originalText}Autor překladu: ${authors[0].name}';
+      } else {
+        return '${originalText}Autoři překladu: ${authors.map((author) => author.name).toList().join(", ")}';
+      }
     }
-
-    return _authorsText ??= text;
   }
 
-  bool get hasExternals => youtubes.isNotEmpty;
+  List<External> get pdfs => externals.where((external) => external.type == MediaType.pdf).toList();
+  List<External> get youtubes => externals.where((external) => external.type == MediaType.youtube).toList();
+  List<External> get mp3s => externals.where((external) => external.type == MediaType.mp3).toList();
 
-  final List<External> externals = [];
-  List<External> get youtubes => externals.where((external) => external.isYoutubeVideo).toList();
+  @override
+  String toString() => 'SongLyric(id: $id, name: $name)';
+
+  @override
+  bool operator ==(Object other) => other is SongLyric && id == other.id;
+
+  @override
+  int get hashCode => id;
 }
