@@ -1,4 +1,6 @@
-import 'package:zpevnik/models/model.dart' as model;
+// ignore: unnecessary_import
+import 'package:objectbox/objectbox.dart';
+import 'package:zpevnik/models/objectbox.g.dart';
 
 enum TagType {
   liturgyPart,
@@ -42,7 +44,59 @@ extension TagTypeExtension on TagType {
     }
   }
 
-  int get rawValue => TagType.values.indexOf(this);
+  static TagType fromRawValue(int rawValue) {
+    switch (rawValue) {
+      case 0:
+        return TagType.liturgyPart;
+      case 1:
+        return TagType.liturgyPeriod;
+      case 2:
+        return TagType.saints;
+      case 3:
+        return TagType.historyPeriod;
+      case 4:
+        return TagType.instrumentation;
+      case 5:
+        return TagType.genre;
+      case 6:
+        return TagType.musicalForm;
+      case 7:
+        return TagType.sacredOccasion;
+      case 8:
+        return TagType.language;
+      case 9:
+        return TagType.generic;
+      default:
+        return TagType.unknown;
+    }
+  }
+
+  int get rawValue {
+    switch (this) {
+      case TagType.liturgyPart:
+        return 0;
+      case TagType.liturgyPeriod:
+        return 1;
+      case TagType.saints:
+        return 2;
+      case TagType.historyPeriod:
+        return 3;
+      case TagType.instrumentation:
+        return 4;
+      case TagType.genre:
+        return 5;
+      case TagType.musicalForm:
+        return 6;
+      case TagType.sacredOccasion:
+        return 7;
+      case TagType.language:
+        return 8;
+      case TagType.generic:
+        return 9;
+      default:
+        return 10;
+    }
+  }
 
   bool get supported {
     switch (this) {
@@ -63,7 +117,7 @@ extension TagTypeExtension on TagType {
       case TagType.generic:
         return 'Příležitosti';
       case TagType.liturgyPart:
-        return 'Liturgie - mše svatá';
+        return 'Mše svatá';
       case TagType.liturgyPeriod:
         return 'Liturgický rok';
       case TagType.saints:
@@ -78,29 +132,46 @@ extension TagTypeExtension on TagType {
   }
 }
 
-// wrapper around Tag db model for easier field access
+@Entity()
 class Tag {
-  final model.Tag entity;
+  @Id(assignable: true)
+  final int id;
 
-  Tag(this.entity);
+  final String name;
 
-  Tag.clone(Tag tag) : entity = tag.entity;
+  final int dbType;
 
-  static Future<List<Tag>> get tags async {
-    final entities = await model.Tag().select().orderBy('id').toList();
+  Tag(this.id, this.name, this.dbType);
 
-    return entities.map((entity) => Tag(entity)).toList();
+  factory Tag.fromJson(Map<String, dynamic> json) {
+    return Tag(
+      int.parse(json['id'] as String),
+      json['name'] as String,
+      TagTypeExtension.fromString(json['type_enum'] as String).rawValue,
+    );
   }
 
-  int get id => entity.id ?? 0;
-  String get name => entity.name ?? '';
-  TagType get type => TagTypeExtension.fromString(entity.type_enum ?? '');
+  static List<Tag> fromMapList(Map<String, dynamic> json) {
+    return (json['tags_enum'] as List).map((json) => Tag.fromJson(json)).toList();
+  }
 
-  // TODO: remove isSelected and use it from songlyricsprovider
-  bool _isSelected = false;
-  bool get isSelected => _isSelected;
+  static List<Tag> load(Store store, _) {
+    final query = store.box<Tag>().query();
+    query.order(Tag_.id);
 
-  void toggleIsSelected() => _isSelected = !_isSelected;
+    return query.build().find();
+  }
+
+  TagType get type => TagTypeExtension.fromRawValue(dbType);
+
+  @override
+  String toString() => 'Tag(id: $id, name: $name)';
+
+  @override
+  bool operator ==(Object other) => other is Tag && id == other.id;
+
+  @override
+  int get hashCode => id;
 }
 
 class TagsSection {
