@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:zpevnik/providers/data.dart';
 import 'package:zpevnik/providers/songbooks.dart';
-import 'package:zpevnik/screens/components/songbooks_grid_view.dart';
+import 'package:zpevnik/components/songbooks_grid_view.dart';
 
 class SongbooksScreen extends StatelessWidget {
   const SongbooksScreen({Key? key}) : super(key: key);
